requirements.txt
torch
torchvision
torchaudio

transformers
sentencepiece
datasets
evaluate
huggingface-hub
tokenizers
accelerate

opencv-python
opencv-contrib-python

numpy
pandas
matplotlib
scikit-learn
scipy

tqdm
Pillow
nltk

rouge
pycocoevalcap

jupyter
ipykernel

jsonschema
pyyaml
